LifeCycle class
